Moved all declarations into separate .h's There are other IDE's other than Arduino with a single *.ino file out there.
